게시글과 댓글의 수정/삭제 권한 부여하기

+++ app/controllers/foods_controller.rb
@@ -65,6 +65,10 @@ class FoodsController < ApplicationController
   end
   def edit
       @post = Post.find(params[:id])
+      if @post.user_id != session[:user_id]
+          flash[:alert] = "수정 권한이 없습니다."
+          redirect_to :back
+      end
   end
 
   def edit_complete
@@ -83,14 +87,24 @@ class FoodsController < ApplicationController
 
   def delete_complete
       post = Post.find(params[:id])
-      post.destroy
-      flash[:alert] = "삭제 되었습니다. "
-      redirect_to "/"
+      if post.user_id == session[:user_id]
+        post.destroy
+        flash[:alert] = "삭제 되었습니다. "
+        redirect_to "/"
+      else
+          flash[:alert] = "삭제 권한이 없습니다."
+          redirect_to :back
+      end
   end
     def delete_comment_complete
         comment = Comment.find(params[:id])
-        comment.destroy
-        flash[:alert] = "댓글이 삭제되었습니다."
-        redirect_to "/foods/show/#{comment.post_id}"
+        if comment.user_id == session[:user_id]
+            comment.destroy
+            flash[:alert] = "댓글이 삭제되었습니다."
+            redirect_to "/foods/show/#{comment.post_id}"
+        else
+            flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+            redirect_to :back
+        end
     end
 end
